Move `Manufacturer` enum to dedicated file
